requirements.txt
Flask==3.0.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
PyPDF2==4.0.1
pdfplumber==0.10.3
python-docx==0.8.11
matplotlib==3.8.2
seaborn==0.13.0
Werkzeug==3.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
